edge_device/src/fastapi_gcp/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
google-cloud-firestore==2.14.0
uvicorn==0.29.0

